Extract graph test in a dedicated file
This is the first step of my quest to benchmark these functions
I just want to take these functions and without thinking at what it's
doing, execute both test and benchmarks.
