+++ massa-consensus/src/tests/mod.rs
@@ -24,4 +24,5 @@ mod scenarios_pruning;
 mod scenarios_reward_split;
 mod scenarios_send_block;
 mod scenarios_wishlist;
+mod test_block_graph;
 pub mod tools;
